+++ src/main/java/com/logandhillon/typeofwar/entity/PlayerObject.java
@@ -38,7 +38,7 @@ public class PlayerObject extends GameObject {
     }
 
     @Override
-    protected void onRender(GraphicsContext g, double x, double y) {
+    protected void onRender(GraphicsContext g, float x, float y) {
         // render skin
         g.setFill(color);
         g.fillRect(x, y - (SIZE / 2.0), SIZE, SIZE);
